requirement.txt
bcrypt==3.1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
django-rest-registration==0.3.13
django-templated-mail==1.1.1
djangorestframework==3.9.0
httpie==1.0.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.19
Pygments==2.2.0
PyJWT==1.4.2
pytz==2018.5
requests==2.20.1
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.
